+++ swift/Sources/LibSignalClient/chat/UnauthKeysService.swift
@@ -9,6 +9,7 @@ import SignalFfi
 public enum UserBasedAuthorization: Sendable {
     case accessKey(Data)
     case groupSend(GroupSendFullToken)
+    case unrestrictedUnauthenticatedAccess
 }
 
 public enum DeviceSpecifier: Sendable {
@@ -77,7 +78,7 @@ extension UnauthenticatedChatConnection: UnauthKeysService {
                 withNativeHandle { chatService in
                     target.withPointerToFixedWidthBinary { target in
                         groupSendFullToken.withUnsafeBorrowedBuffer { auth in
-                            signal_unauthenticated_chat_connection_get_pre_keys_access_group_auth(
+                            signal_unauthenticated_chat_connection_get_pre_keys_group_auth(
                                 promise,
                                 tokioAsyncContext.const(),
                                 chatService.const(),
@@ -86,6 +87,22 @@ extension UnauthenticatedChatConnection: UnauthKeysService {
                                 device,
                             )
                         }
+                    }
+                }
+            }
+        case .unrestrictedUnauthenticatedAccess:
+            out = try await self.tokioAsyncContext.invokeAsyncFunction {
+                promise,
+                tokioAsyncContext in
+                withNativeHandle { chatService in
+                    target.withPointerToFixedWidthBinary { target in
+                        signal_unauthenticated_chat_connection_get_pre_keys_unrestricted_auth(
+                            promise,
+                            tokioAsyncContext.const(),
+                            chatService.const(),
+                            target,
+                            device,
+                        )
                     }
                 }
             }